Added notes and plans for the wave system & spawning
Might need more scripts or ref from others like the enemy scripts. I can do all that tomorrow to make a playable prototype and see if that works with the other scripts. Also, might revamp the spawning so that we don't have to create objects on the go. Rather have them on the map ready to go, just off-screen. I heard this improves performance. I mostly experimented on another separate unity project to keep things safe on the main and pass things that work.