Finally have a working spawn
It's simple and easy; it spawns enemies every 5 seconds at a spawn point that randomizes the range they spawn. The only way to stop the spawn is for the player to die. I was thinking for saving time maybe we can cap that number spawned and if the player kills a certain amount of enemies that round then move on to the next scene/round.