Worked on HealthUI script and spawning.
It works but the Health-UI needs more work. The prefabs of enemies are not taking in images of blood splatter and hurt images but deal damages still. You just won't know. Also added in a line if Player dies Quit.Application that works but we can change it to whatever; restart for example.