Added Sphere and others
Used two scripts for the player to take the damage from the enemy, HealthController which is attached to an empty object on the scene of the game placed anywhere, just have to drag the bloodspatter images. Let me know when you add the enemy I'll tell you where the enemy sphere is located on the body.